Let a MoonLive function return a value, and make Aurora volumetric
A script's helper can now compute rather than only act: `int level() { ... }`
and the call is an expression. The MoonLive Aurora uses it to sample a real 3D
field, so on a cube the curtains stand through the volume instead of one slice
repeating. Three review findings on the polar and noise kernels are closed.

Performance: desktop unchanged (Aurora 64x64 p50 803us, within noise of 786us
before); no ESP32 measurement, since no tick-path code changed.

**Core**
- Every backend's `callLabel` preserves the whole vreg pool across a script call
  and delivers the callee's value into the destination register. Xtensa's window
  rotation is why the result arrives in the caller's a10: call8 rotates by 8, so
  the callee's a2 is this frame's a10, and it is parked in the frame before the
  pool restore runs over it.
- `writesDst` consults `CallScript`'s `b` flag, so a value-returning call is seen
  as DEFINING its destination. It was not, and the register allocator remapped
  every source while leaving the dst unmapped: after a compaction the call wrote
  its pre-compaction register and the consumer read the new one. Only reachable
  at a squeezed budget, which is the normal case on a ten-register Xtensa, so the
  host suite never saw it and hardware would have.
- `kMaxLocals` is 32: a volumetric script needs 19 slots, and the cost is stack
  (148 bytes of frame), never the encoding.

**Light domain**
- `warpImpl` adds its displacement in unsigned arithmetic. Casting the coordinate
  to int32_t first was signed overflow for any coordinate past 2^31, which
  `r * zoom + drift` reaches on a large fixture.
- NoiseEffect's morph passes time alone on the third axis. With a depth term it
  was a second drift, and the catalog card promised otherwise.
- `PolarLut::addressOf` is the one home for the projections: the table builder
  calls it instead of holding a second copy, the computed fallback keeps the
  selected mapping in Aurora, PolarNoise and Tunnel (it computed cylindrical
  unconditionally, so a device short of memory silently changed composition), and
  `mappingOf` replaces the clamp's four copies.

**Tests**
- A returned value survives a squeezed budget, entered at the script's `tick`
  rather than the block start, plus the caller-side result stash pinned per device
  ISA. Nothing else reaches those two backends, where a wrong encoding is silent.
- The computed fallback matches the table under all three mappings.
- Morph shows the same field in every slice; FireEffect gains the golden it lacked.

**Docs**
- The spec said a function returns nothing; the builtin table lacked `fbm3`,
  `warp3` and `setPaletteColorZ`; the Aurora card said 2D while the code is D3.
- Phase 2's 16-bit Layer is recorded as built, measured and stashed: a wide frame
  buffer costs +56.5% on the S3 and +57.9% on desktop, the two agreeing. Phases 3
  and 4 start in 8 bits. Migrating a raw-byte effect onto `draw::pixel` is a
  measured regression (+170%, +327% for the flat-index form), so that step is
  withdrawn.

+++ src/core/noise.h
@@ -385,10 +385,14 @@ inline uint8_t warpImpl(uint32_t x, uint32_t y, uint32_t z, uint16_t strength, u
         dx = (static_cast<int32_t>(inoise8(x, y)) - 128) * strength / 128;
         dy = (static_cast<int32_t>(inoise8(x + 0x9E37u, y + 0x7C15u)) - 128) * strength / 128;
     }
-    const uint32_t sx = static_cast<uint32_t>(static_cast<int32_t>(x) + dx);
-    const uint32_t sy = static_cast<uint32_t>(static_cast<int32_t>(y) + dy);
+    // The displacement is added in UNSIGNED arithmetic. Casting the coordinate to int32_t first
+    // was signed overflow (undefined behavior) for any coordinate past 2^31, which a scaled field
+    // reaches easily: `r * zoom + drift` on a large fixture is already past it. Unsigned wrapping is
+    // defined, and wrapping is what a noise coordinate wants anyway.
+    const uint32_t sx = x + static_cast<uint32_t>(dx);
+    const uint32_t sy = y + static_cast<uint32_t>(dy);
     if constexpr (Dims > 2) {
-        return fbm8(sx, sy, static_cast<uint32_t>(static_cast<int32_t>(z) + dz), octaves);
+        return fbm8(sx, sy, z + static_cast<uint32_t>(dz), octaves);
     } else {
         return fbm8(sx, sy, octaves);
     }
